Update contact.html
I loved the design of this page! When I change the view for a smaller screen, text, and social media are in the same row and everything else appears in their own row. That is the only change I would suggest. I am adding a comment so I can propose a change.